enhance unit tests
